Updated file structure and fixed imports
