+++ backend/src/main/java/cs/superleague/shopping/requests/GetItemsRequest.java
@@ -1,13 +1,14 @@
-package shopping.requests;
+package cs.superleague.shopping.requests;
 
 import java.util.UUID;
 
-public class GetStoreOpenRequest {
+public class GetItemsRequest {
+
     UUID storeID;
 
-    public GetStoreOpenRequest(UUID storeID)
+    public GetItemsRequest(UUID storeID)
     {
-        this.storeID=storeID;
+        this.storeID= storeID;
     }
 
     public void setStoreID(UUID storeID) {
