+++ backend/src/main/java/cs/superleague/shopping/responses/GetStoreOpenResponse.java
@@ -1,4 +1,4 @@
-package shopping.responses;
+package cs.superleague.shopping.responses;
 
 import java.util.Date;
